correction ecies receiver private (secret) key
The ECIES reveiver's public key is a random scalar. Then, it is also the secret to be split.
Previous version: wrongly took ECIES reveiver's public key as the y-coordinate of the secret share. So, a point p=(x,y) was encrypted with a secret key k which was also the y-coordinate of the point. That's to say,
E(k, p) where p=(x,y) and k=y